Added ListBlocks
Added ListBlocks and did some refactoring to be able to reuse code
between the three block types.

+++ src/Taskling.SqlServer.Database/Tables/DateRangeBlockExecution.sql
@@ -1,7 +1,7 @@
 ﻿CREATE TABLE [Taskling].[DateRangeBlockExecution](
 	[DateRangeBlockExecutionId] [bigint] IDENTITY(1,1) NOT NULL,
 	[TaskExecutionId] [int] NOT NULL,
-	[DateRangeBlockId] [tinyint] NOT NULL,
+	[DateRangeBlockId] [bigint] NOT NULL,
 	[StartedAt] [datetime] NOT NULL,
 	[CompletedAt] [datetime] NULL,
 	[BlockExecutionStatus] [tinyint] NOT NULL,
